backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
qrcode==7.4.2
pandas==2.1.3
openpyxl==3.1.2